requirements.txt
numpy
pandas
pyyaml
matplotlib
scikit-learn
torch
timm
pyts
PyWavelets
tqdm
pyarrow